Refactor: Moved models from domain to repository
