+++ app/src/main/java/com/example/pokedexdesafio/data/repository/GetPokemonDetailRepoImpl.kt
@@ -1,7 +1,7 @@
 package com.example.pokedexdesafio.data.repository
 
 import com.example.pokedexdesafio.data.services.GetPokemonDetailService
-import com.example.pokedexdesafio.domain.model.PokemonDetail
+import com.example.pokedexdesafio.data.model.PokemonDetail
 import com.example.pokedexdesafio.domain.repository.GetPokemonDetailsRepo
 import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.schedulers.Schedulers
