+++ app/src/main/java/com/example/pokedexdesafio/domain/repository/GetPokemonRepo.kt
@@ -1,6 +1,6 @@
 package com.example.pokedexdesafio.domain.repository
 
-import com.example.pokedexdesafio.domain.model.PokemonResponse
+import com.example.pokedexdesafio.data.model.PokemonResponse
 import io.reactivex.rxjava3.core.Single
 import retrofit2.Response
