+++ app/src/main/java/com/example/pokedexdesafio/presentation/activity/PokemonDetailActivity.kt
@@ -7,7 +7,7 @@ import com.bumptech.glide.Glide
 import com.example.pokedexdesafio.R
 import com.example.pokedexdesafio.core.platform.BaseActivity
 import com.example.pokedexdesafio.core.utils.KEY_ID
-import com.example.pokedexdesafio.domain.model.PokemonDetail
+import com.example.pokedexdesafio.data.model.PokemonDetail
 import com.example.pokedexdesafio.presentation.viewmodel.PokemonDetailViewModel
 import retrofit2.Response
 import javax.inject.Inject
